pyproject.toml
[tool.poetry]
name = "ke-sh-crawler"
version = "0.1.0"
description = ""
authors = ["Charles Jing <cjing@mintel.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Scrapy = "^2.6.1"
scrapy-fake-useragent = "^1.4.4"
psycopg2 = "^2.9.3"
SQLAlchemy = "^1.4.39"
sentry-sdk = "^1.7.2"
tqdm = "^4.64.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
